+++ server/src/datastore.js
@@ -10,6 +10,7 @@ export const setPlayerLanguage = (...a) => impl.setPlayerLanguage(...a);
 export const getCurrentLevel = (...a) => impl.getCurrentLevel(...a);
 export const advanceProgress = (...a) => impl.advanceProgress(...a);
 export const recordGame = (...a) => impl.recordGame(...a);
+export const recordEvent = (...a) => impl.recordEvent(...a);
 export const getPlayerGames = (...a) => impl.getPlayerGames(...a);
 export const stats = (...a) => impl.stats(...a);
 export const getShips = (...a) => impl.getShips(...a);